+++ buildpacks/npm/lib/build.sh
@@ -16,6 +16,24 @@ source "$bp_dir/lib/utils/json.sh"
 # shellcheck source=/dev/null
 source "$bp_dir/lib/utils/log.sh"
 
+fail_multiple_lockfiles() {
+	local build_dir=$1
+	local has_modern_lockfile=false
+	if [[ -f "$build_dir/package-lock.json" || -f "$build_dir/yarn.lock" ]]; then
+		has_modern_lockfile=true
+	fi
+
+	if [[ -f "$build_dir/package-lock.json" && -f "$build_dir/yarn.lock" ]]; then
+		error "Build failed because two different lockfiles were detected: package-lock.json and yarn.lock"
+		exit 1
+	fi
+
+	if [[ $has_modern_lockfile && -f "$build_dir/npm-shrinkwrap.json" ]]; then
+		error "Build failed because multiple lockfiles were detected"
+		exit 1
+	fi
+}
+
 clear_cache_on_stack_change() {
 	local layers_dir=$1