+++ buildpacks/yarn/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add license to buildpack.toml ([#17](https://github.com/heroku/buildpacks-node/pull/17))
 - Flush cache when stack image changes ([#28](https://github.com/heroku/buildpacks-node/pull/28))
 - Trim whitespace when getting stack name ([#29](https://github.com/heroku/buildpacks-node/pull/29))
+- Fail if two lock files are detected ([#30](https://github.com/heroku/buildpacks-node/pull/30))
 
 ## [0.1.1] 2021/01/20